refactor: Explode TransferPage components into files
