+++ src/ducks/transfers/utils.js
@@ -1,0 +1,2 @@
+export const isLoginFailed = error =>
+  error.message && error.message.includes('LOGIN_FAILED')
